Add updated metrics
